Re-organizing the backend
