+++ .gitignore
@@ -6,3 +6,4 @@ tmp/
 *.so.*
 *.o
 *.la
+src/ega_ssh_keys
