+++ src/Makefile
@@ -7,22 +7,27 @@
 NSS_LD_SONAME=-Wl,-soname,libnss_ega.so.2
 NSS_LIBRARY=libnss_ega.so.2.0
 PAM_LIBRARY = pam_ega.so
+KEYS_EXEC = ega_ssh_keys
 
 CC=gcc
 LD=ld
 AS=gcc -c
-CFLAGS=-Wall -Wstrict-prototypes -Werror -fPIC -O2 -I. -I$(shell pg_config --includedir)
-LIBS=-lpq -lpam -lcurl -ljq -L$(shell pg_config --libdir)
+CFLAGS=-Wall -Wstrict-prototypes -Werror -fPIC -O2 -I.
+LIBS=-lpam -lcurl -ljq
 
 LIBDIR=/usr/local/lib/ega
+BINDIR=/usr/local/bin
 
 HEADERS = debug.h config.h backend.h cega.h homedir.h $(wildcard blowfish/*.h)
 
-NSS_SOURCES = nss.c config.c backend_fs.c cega.c homedir.c
+NSS_SOURCES = nss.c config.c backend.c cega.c homedir.c
 NSS_OBJECTS = $(NSS_SOURCES:%.c=%.o)
 
-PAM_SOURCES = pam.c storage.c config.c backend_fs.c cega.c homedir.c $(wildcard blowfish/*.c)
+PAM_SOURCES = pam.c config.c backend.c cega.c homedir.c $(wildcard blowfish/*.c)
 PAM_OBJECTS = $(PAM_SOURCES:%.c=%.o) blowfish/x86.o
+
+KEYS_SOURCES = keys.c config.c backend.c
+KEYS_OBJECTS = $(KEYS_SOURCES:%.c=%.o)
 
 .PHONY: all debug clean install install-nss install-pam
 .SUFFIXES: .c .o .S .so .so.2 .so.2.0
@@ -39,6 +44,10 @@ $(NSS_LIBRARY): $(HEADERS) $(NSS_OBJECTS)
 $(PAM_LIBRARY): $(HEADERS) $(PAM_OBJECTS)
 	@echo "Linking objects into $@"
 	@$(LD) -x --shared -o $@ $(LIBS) $(PAM_OBJECTS)
+
+$(KEYS_EXEC): $(HEADERS) $(KEYS_OBJECTS)
+	@echo "Creating $@"
+	@$(CC) -o $@ $(KEYS_OBJECTS)
 
 blowfish/x86.o: blowfish/x86.S $(HEADERS)
 	@echo "Compiling $<"
@@ -58,11 +67,17 @@ install-pam: $(PAM_LIBRARY)
 	@echo "Installing $< into $(LIBDIR)"
 	@install $< $(LIBDIR)
 
-install: install-nss install-pam
+install-keys: $(KEYS_EXEC)
+	@[ -d $(BINDIR) ] || { echo "Creating bin dir: $(BINDIR)"; install -d $(BINDIR); }
+	@echo "Installing $< into $(BINDIR)"
+	@install $< $(BINDIR)
+
+install: install-nss install-pam install-keys
 	@echo "Do not forget to run ldconfig and create/configure the file /etc/ega/auth.conf"
 	@echo "Look at the auth.conf.sample here, for example"
 
 
 clean:
-	-rm -f $(NSS_LIBRARY) $(NSS_OBJECTS) backend*.o
-	-rm -f $(PAM_LIBRARY) $(PAM_OBJECTS) backend*.o
+	-rm -f $(NSS_LIBRARY) $(NSS_OBJECTS)
+	-rm -f $(PAM_LIBRARY) $(PAM_OBJECTS)
+	-rm -f $(KEYS_EXEC) $(KEYS_OBJECTS)
